dagster/requirements.txt
alembic==1.13.2
aniso8601==9.0.1
annotated-types==0.7.0
ansicolors==1.1.8
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
backoff==2.2.1
beautifulsoup4==4.12.3
bitmath==1.3.3.1
bleach==6.1.0
boto3==1.34.146
botocore==1.34.146
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
coloredlogs==14.0
comm==0.2.2
croniter==2.0.7
cryptography==43.0.0
dagster==1.7.15
dagster-aws==0.23.15
dagster-cloud==1.7.15
dagster-cloud-cli==1.7.15
dagster-graphql==1.7.15
dagster-k8s==0.23.15
dagster-pipes==1.7.15
dagster-postgres==0.23.15
dagster-webserver==1.7.15
dagstermill==0.23.15
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
dnspython==2.6.1
docstring_parser==0.16
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.15.4
fqdn==1.5.1
fsspec==2023.12.2
github3.py==4.0.1
google-auth==2.32.0
gql==3.5.0
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
grpcio==1.64.1
grpcio-health-checking==1.62.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.3
humanfriendly==10.0
idna==3.7
ipykernel==6.29.5
ipython==8.18.0
ipython-genutils==0.2.0
ipywidgets==8.1.3
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==7.4.9
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
jwt==1.3.1
kaggle==1.6.14
kubernetes==30.1.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
minio==7.2.7
mistune==3.0.2
mmhash3==3.0.1
multidict==6.0.5
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
noteable-origami==0.0.35
notebook==7.2.1
notebook_shim==0.2.4
numpy==2.0.1
oauthlib==3.2.2
orjson==3.10.6
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
papermill==2.6.0
papermill-origami==0.0.30
parso==0.8.4
pendulum==3.0.0
pex==2.11.0
pexpect==4.9.0
platformdirs==4.2.2
polars==1.2.1
prometheus_client==0.20.0
prompt-toolkit==3.0.36
protobuf==4.25.3
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.7
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==1.10.17
pydantic_core==2.20.1
pydeequ==1.4.0
Pygments==2.18.0
PyJWT==2.8.0
pymongo==4.8.0
pyparsing==3.1.2
pyspark==3.5.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
questionary==2.0.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.19.1
rsa==4.9
s3transfer==0.10.2
scrapbook==0.5.0
Send2Trash==1.8.3
sending==0.3.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.31
stack-data==0.6.3
starlette==0.38.0
strictyaml==1.7.3
structlog==22.3.0
tabulate==0.9.0
tenacity==8.5.0
terminado==0.18.1
text-unidecode==1.3
time-machine==2.14.2
tinycss2==1.3.0
tomli==2.0.1
toposort==1.10
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
universal_pathlib==0.2.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.3
uvloop==0.19.0
watchdog==4.0.1
watchfiles==0.22.0
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
widgetsnbextension==4.0.11
yarl==1.9.4
